requirements.txt
Pillow
numpy
PyMuPDF
PyPDF2
reportlab